Update to work with modern MongoDB
I'm a total noob at Common Lisp and MongoDB. I just happened to be working through https://github.com/adamtornhill/LispForTheWeb by @adamtornhill, and managed to hack a bit on this to get it working. I've had literally zero looks at the rest of the code, so I'm sure there's much more work to be done. At least this is a start, and hopefully not completely wrong.